refactor: merge `import-sync.ts` into `import.ts`
